made Course.number non-null
Design Design Rationale : Every table has an auto-generated auto-incremented integer primary key because 
a. It is best practice for PKs to be as small and efficient as possible.
b. From a security-standpoint, it is best to have a primary key that the user cannot directly see. Instead, the user handles only a surrogate key (such as a unique netID), while the real PK is hidden.